#12 & #14 Reorganize files for beepboop deployment
