+++ Dockerfile
@@ -5,6 +5,8 @@ WORKDIR /usr/src/app
 
 COPY app.js /usr/src/app/
 COPY dbmanager.js /usr/src/app/
+COPY takeOrder.js /usr/src/app/
+COPY witlogic.js /usr/src/app/
 COPY package.json /usr/src/app/
 RUN npm install
