+++ README.md
@@ -1,4 +1,4 @@
-# goferbot
+# goferbot <a href="https://slack.com/oauth/authorize?scope=bot&client_id=21451526290.39930048497"><img alt="Add to Slack" height="40" width="139" src="https://platform.slack-edge.com/img/add_to_slack.png" srcset="https://platform.slack-edge.com/img/add_to_slack.png 1x, https://platform.slack-edge.com/img/add_to_slack@2x.png 2x" /></a>
 ### A slack bot to coordinate the morning coffee run for small offices
 
 Goferbot uses natural language processing to collect orders for the morning coffee run.
