--- takeOrder.spec.js
@@ -1,10 +1,0 @@
-var assert = require('chai').assert;
-var takeOrder = require('./takeOrder.js');
-// i'll need an entire mock convo to pipe through
-// including all it's methods >__<
-
-describe('Take an order', function() {
-  it('Should prompt for an order', function(done) {
-    
-  });
-});
